Move logic to class
